+++ invenio_pidstore_extra/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# Copyright (C) 2024 University of Münster.
+# Copyright (C) 2025 University of Münster.
 #
 # invenio-pidstore-extra is free software; you can redistribute it and/or
 # modify it under the terms of the MIT License; see LICENSE file for more
@@ -10,6 +10,6 @@
 
 from .ext import InvenioUrnProvider
 
-__version__ = "0.1.0"
+__version__ = "0.2.0"
 
 __all__ = ("__version__", "InvenioUrnProvider")