+++ invenio_pidstore_extra/services/components/__init__.py
@@ -7,3 +7,7 @@
 # details.
 
 """DNB wrapper."""
+
+from .pids import URNRelationsComponent
+
+__all__ = URNRelationsComponent