+++ CHANGES.rst
@@ -1,5 +1,5 @@
 ..
-    Copyright (C) 2024 University of Münster.
+    Copyright (C) 2024-2025 University of Münster.
 
     invenio-urn-provider is free software; you can redistribute it and/or
     modify it under the terms of the MIT License; see LICENSE file for more
@@ -7,6 +7,12 @@
 
 Changes
 =======
+
+Version 0.2.0
+
+- fully integrated the rest wrapper to DNB urn service
+- adapted to InvenioRDM v13
+- added a new component to set a successor of an URN if a new version has been published
 
 Version 0.1.0 (released TBD)